Tests: move unit tests to `tests\Unit` subdirectory
